upload new folders

+++ Server/src/main/java/com/threeracha/gaewoonhae/utils/jwt/AuthTokensGenerator.java
@@ -14,7 +14,7 @@ public class AuthTokensGenerator {
 
     private final JwtTokenProvider jwtTokenProvider;
 
-    public AuthTokens generate(Integer userId) {
+    public AuthTokens generate(Long userId) {
         long now = (new Date()).getTime();
         Date accessTokenExpiredAt = new Date(now + ACCESS_TOKEN_EXPIRE_TIME);
         Date refreshTokenExpiredAt = new Date(now + REFRESH_TOKEN_EXPIRE_TIME);
